fleshes out CreateRockTheVotePostInRogue job
